Add the function to generate top3 node and link
In draw_skeleton2D.py add the top 3 nodes and top 3 link function that
can find the top3 node and link based on the SMPL format and then
generate the json file according to the video.

Add the read_config function in draw_skeleton2D.py to read the setting
of attention_node, attention_matrix, output directory and the path of
dataset.

Simplify the way to find the top 3 in draw function by using the
dictionary data structure.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>