Fix AlphaPose and attention node frame mismatch
Fix the number of frame inconsistency between AlphaPose predictions and
the attention node. The index of the frame predicted by AlphaPose should
be scaled by dividing the length of the attention node by the number of
frames predicted by AlphaPose.

index * (length of attention node) / (# of frame predicted by alphapose)

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>