Add handling for inconsistent filename of ErrorSeg
In file `draw_skeleton2D` and `GPTprepare`, add the file list that the
error segment setting do not use.

In file `datasetError`, add the implemenation of taking the
`feature_start_frame` and `feature_end_frame` as the start frame and end
frame when clipping video.

Add the file `filepackage` to pack the visualization of golden_file that
the predicted alphapose is correct.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>